pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.5.5",
    "fastapi>=0.121.2",
    "ipykernel>=7.1.0",
    "langchain>=0.2.14",
    "langchain-community>=0.2.0",
    "langchain-openai>=0.1.22",
    "langchain-text-splitters>=0.2.0",
    "openai>=1.51.0",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=5.1.2",
    "tiktoken>=0.7.0",
    "uvicorn[standard]>=0.38.0",
]
